reorganizacja kodu, ułatwienie zmieniania parametrów dla poszczególnych gatunków
