+++ src/generowanie_muzyki/R_B.py
@@ -1,0 +1,5 @@
+from default import *
+
+class PieceRB(Piece):
+    def __init__(this, metrum, skala, tempo):
+        Piece.__init__(this, metrum, skala, tempo, [8, 33, 34, 84, 89], [35, 36, 39, 60, 73])
