+++ src/generowanie_muzyki/pop.py
@@ -1,0 +1,5 @@
+from default import *
+
+class PiecePop(Piece):
+    def __init__(this, metrum, skala, tempo):
+        Piece.__init__(this, metrum, skala, tempo, [5, 6, 26, 28, 29, 33, 49, 50, 81, 87, 89, 90, 96, 97], [35, 36, 41, 48, 53, 56])
