+++ src/generowanie_muzyki/test_muzyka.py
@@ -1,4 +1,4 @@
-from muzyka import *
+from default import *
 
 
 class TestDzwiek:
